refactor: move handles into modules
